Define 'length' as a signed variable, so it can handle an error value of -1 (v5->v4 patch merge)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V4-2-patches@8306 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_api.c
@@ -4000,7 +4000,7 @@ _sess_read(void *sessp,
     u_char packet[PACKET_LENGTH], *packetptr = packet;
     snmp_ipaddr        from;
     struct sockaddr_in *fromIp = (struct sockaddr_in *)&from;
-    size_t length = 0;
+    int length = 0;
     struct snmp_pdu *pdu;
     struct request_list *rp, *orp = NULL;
     int ret = 0, handled = 0;
